Jakarta Support
This change updates the project to support both the javax and jakarta
namespaces for JEE projects.  It does this by creating two separate
Maven projects, one for each, which are then shaded into a single JAR
used by everyone.  The contents of META-INF/services and
META-INF/spring.factories route to the correct implementation preventing
class loading problems.
