+++ .gitignore
@@ -1,4 +1,5 @@
 .idea/libraries
 .idea/tasks.xml
 .idea/workspace.xml
+dependency-reduced-pom.xml
 target
