+++ .idea/encodings.xml
@@ -2,5 +2,13 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper-jakarta/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper-jakarta/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper-javax/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper-javax/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/java-buildpack-client-certificate-mapper/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
   </component>
 </project>
