+++ .java-version
@@ -1,0 +1,1 @@
+17.0
