+++ java-buildpack-client-certificate-mapper/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.cloudfoundry.router.jakarta.ClientCertificateMapperAutoConfiguration,\
+org.cloudfoundry.router.javax.ClientCertificateMapperAutoConfiguration
